Extract pending poll handling into a separate module
