+++ src/worker_process.rs
@@ -60,6 +60,7 @@ use crate::worker_supervisor::{
 
 mod control_prefix;
 mod output_state;
+mod pending_poll;
 mod write_dispatch;
 mod write_preflight;
 
@@ -1175,231 +1176,6 @@ impl WorkerManager {
         }
         let end_offset = self.output.end_offset().unwrap_or(0);
         Some(ReplyWithOffset { reply, end_offset })
-    }
-
-    /// Serves empty-input polls and busy follow-up replies for a timed-out request.
-    /// Each poll only returns newly available output, but the server may keep appending it to one transcript file.
-    fn poll_pending_output_files(
-        &mut self,
-        timeout: Duration,
-    ) -> Result<ReplyWithOffset, WorkerError> {
-        let poll_start = std::time::Instant::now();
-        let mut timed_out = false;
-        let mut completed_request = false;
-        let mut consumed_completion = false;
-        let mut completion = CompletionInfo::empty();
-
-        if let Some(err) = self.settled_pending_error.take() {
-            let _ = self.reset_preserving_detached_prefix_item_count();
-            return Err(err);
-        }
-        if self.pending_request {
-            match self.wait_for_request_completion(timeout) {
-                Ok(info) => {
-                    self.clear_pending_request_state();
-                    if info.session_end_seen {
-                        self.note_session_end(true);
-                    }
-                    completion = info;
-                    completed_request = true;
-                    consumed_completion = true;
-                }
-                Err(WorkerError::Timeout(_)) => {
-                    let worker_exited = match self.process.as_mut() {
-                        Some(process) => !process.is_running()?,
-                        None => true,
-                    };
-                    if worker_exited {
-                        self.note_session_end(true);
-                        self.clear_pending_request_state();
-                        completion.session_end_seen = true;
-                        completed_request = true;
-                        consumed_completion = true;
-                    } else {
-                        timed_out = true;
-                    }
-                }
-                Err(err) => return Err(err),
-            }
-        }
-        if !timed_out
-            && !completed_request
-            && let Some(info) = self.settled_pending_completion.take()
-        {
-            if info.session_end_seen {
-                self.note_session_end(false);
-            }
-            completion = info;
-            consumed_completion = true;
-        }
-        let fallback_input = if !timed_out && consumed_completion {
-            self.take_input_fallback(&completion)
-        } else {
-            InputFallback::default()
-        };
-        if !timed_out && consumed_completion {
-            self.wait_for_late_files_output_after_settled_completion(timeout);
-        }
-
-        let FormattedPendingOutput {
-            mut contents,
-            saw_stderr,
-        } = if timed_out {
-            self.drain_formatted_output()
-        } else {
-            self.drain_final_formatted_output()
-        };
-        let is_error = saw_stderr;
-
-        if timed_out {
-            let elapsed = self
-                .pending_request_started_at
-                .map(|start| start.elapsed())
-                .unwrap_or_else(|| poll_start.elapsed());
-            contents.push(timeout_status_content(elapsed));
-            return Ok(build_timeout_reply(contents, is_error, 0));
-        }
-
-        let session_end = completion.session_end_seen;
-        let built = build_completed_reply(
-            contents,
-            is_error,
-            0,
-            &completion,
-            session_end,
-            CompletionReplyMode::Files {
-                fallback_input,
-                idle_status_if_empty: true,
-            },
-            self.backend,
-        );
-        self.remember_prompt(built.prompt_to_remember.clone());
-        Ok(built.reply)
-    }
-
-    fn poll_pending_output_pager(
-        &mut self,
-        timeout: Duration,
-        page_bytes: u64,
-    ) -> Result<ReplyWithOffset, WorkerError> {
-        let poll_start = std::time::Instant::now();
-        let start_offset = self.output.current_offset().unwrap_or(0);
-        let mut end_offset = self.output.end_offset().unwrap_or(start_offset);
-        let mut timed_out = false;
-        let mut completed_request = false;
-        let mut completion = CompletionInfo::empty();
-
-        if let Some(err) = self.settled_pending_error.take() {
-            let preserve_pager = self.pager.is_active();
-            let _ = self.reset_with_pager_preserving_detached_prefix_item_count(preserve_pager);
-            return Err(err);
-        }
-        if self.pending_request {
-            match self.wait_for_request_completion(timeout) {
-                Ok(info) => {
-                    self.clear_pending_request_state();
-                    if info.session_end_seen {
-                        self.note_session_end(true);
-                    }
-                    completion = info;
-                    completed_request = true;
-                    end_offset = self.output.end_offset().unwrap_or(end_offset);
-                }
-                Err(WorkerError::Timeout(_)) => {
-                    end_offset = self.output.end_offset().unwrap_or(end_offset);
-                    let worker_exited = match self.process.as_mut() {
-                        Some(process) => !process.is_running()?,
-                        None => true,
-                    };
-                    if worker_exited {
-                        self.note_session_end(true);
-                        self.clear_pending_request_state();
-                        completion.session_end_seen = true;
-                        completed_request = true;
-                    } else {
-                        timed_out = true;
-                    }
-                }
-                Err(err) => return Err(err),
-            }
-        }
-        if !completed_request && let Some(info) = self.settled_pending_completion.take() {
-            if info.session_end_seen {
-                self.note_session_end(false);
-            }
-            completion = info;
-            completed_request = true;
-            end_offset = self.output.end_offset().unwrap_or(end_offset);
-        }
-
-        if end_offset < start_offset {
-            end_offset = start_offset;
-        }
-
-        let (saw_stderr, snapshot) = if completed_request {
-            let completed = snapshot_after_completion(
-                &self.output,
-                start_offset,
-                end_offset,
-                page_bytes,
-                &completion.echo_events,
-                completion.prompt_variants.as_deref(),
-            );
-            (completed.saw_stderr, completed.snapshot)
-        } else {
-            let saw_stderr = self
-                .output
-                .saw_stderr_in_range(start_offset.min(end_offset), end_offset);
-            let snapshot = snapshot_page_with_images(&self.output, end_offset, page_bytes);
-            (saw_stderr, snapshot)
-        };
-        let is_error = saw_stderr;
-        let page_is_error = saw_stderr;
-        let SnapshotWithImages {
-            mut contents,
-            pages_left,
-            buffer,
-            last_range,
-        } = snapshot;
-
-        if timed_out {
-            let elapsed = self
-                .pending_request_started_at
-                .map(|start| start.elapsed())
-                .unwrap_or_else(|| poll_start.elapsed());
-            contents.push(timeout_status_content(elapsed));
-        }
-        pager::maybe_activate_and_append_footer(
-            &mut self.pager,
-            &mut contents,
-            pages_left,
-            page_is_error,
-            buffer,
-            last_range,
-        );
-
-        if timed_out {
-            return Ok(build_timeout_reply(contents, is_error, end_offset));
-        }
-
-        let session_end = completion.session_end_seen;
-        let built = build_completed_reply(
-            contents,
-            is_error,
-            end_offset,
-            &completion,
-            session_end,
-            CompletionReplyMode::Pager {
-                pager_active: self.pager.is_active(),
-                fallback_input_transcript: None,
-            },
-            self.backend,
-        );
-        self.remember_prompt(built.prompt_to_remember.clone());
-        if let Some(pager_prompt) = built.pager_prompt {
-            self.pager_prompt = pager_prompt;
-        }
-        Ok(built.reply)
     }
 
     fn send_worker_request(
